pyproject.toml
[project]
name = "barks-compleat-reader"
# No version required - see '_version.py'
dynamic = ["version"]
description = "Kivy app to navigate and read the Fantagraphics Barks library."
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "better-exceptions==0.3.3",
    "certifi==2025.8.3",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "cryptography>=46.0.3",
    "cython==3.1.4 ; sys_platform == 'darwin'",
    "distro>=1.9.0",
    "docutils==0.22",
    "docutils-stubs==0.0.22",
    "dotenv>=0.9.9",
    "filetype==1.2.0",
    "idna==3.10",
    "intspan==1.6.1",
    "kivy==2.3.1",
    "kivy-deps-angle==0.4.0 ; sys_platform == 'win32'",
    "kivy-deps-glew==0.3.1 ; sys_platform == 'win32'",
    "kivy-deps-sdl2==0.8.0 ; sys_platform == 'win32'",
    "kivy-garden==0.1.5",
    "loguru==0.7.3",
    "loguru-config==0.1.0",
    "numpy==2.2.6",
    "pandas>=2.2.3",
    "pillow==11.3.0",
    "psutil>=7.1.3",
    "pygments==2.19.2",
    "pyobjc-core==11.1 ; sys_platform == 'darwin'",
    "pyobjc-framework-cocoa==11.1 ; sys_platform == 'darwin'",
    "pyphen>=0.17.2",
    "pypiwin32==223 ; sys_platform == 'win32'",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyuca>=1.2",
    "pywin32==311 ; sys_platform == 'win32'",
    "pyyaml==6.0.2",
    "screeninfo==0.8.1",
    "simplemma>=1.1.2",
    "six==1.17.0",
    "typer>=0.21.1",
    "types-docutils==0.22.0.20250822",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "whoosh-reloaded>=2.7.5",
    "win32-setctime==1.2.0 ; sys_platform == 'win32'",
]

[dependency-groups]
dev = [
    "import-linter>=2.9",
    "pyarmor>=9.2.0",
    "pycrucible>=0.3.0",
    "pyinstrument>=5.1.1",
    "pyspellchecker>=0.8.4",
    "pytest>=8.4.2",
    "pytest-benchmark>=5.2.3",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "snakeviz>=2.2.2",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
addopts = """
    --benchmark-time-unit=s
    --benchmark-columns=min,max,mean,stddev
    --benchmark-warmup=on
    --benchmark-group-by=func
"""

[tool.coverage.run]
# Only measure code in your source folder (ignore virtualenv stuff)
source = ["src/barks-reader/src/barks_reader"]

[tool.coverage.report]
# fail_under = 80
show_missing = true
skip_empty = true
# Don't check the tests themselves, only the app code
omit = [
    "*/tests/*",
    "*/__init__.py",
]
